Two-Sum
Used an unordered_map to keep track of the numbers in nums, along with their index. Unordered map is used to allow constant time complexity for lookups and insertions on average.  We iterate the entire nums array. For each number in num[i] we calculate the complement by taking Target - nums[i]; If we find the complement number in the numbermap we return the nums[i] and the complement that adds up to i. If the compliment is not found store the current number and index into the number map. If none we return an empty set.