IsPalindrome (int)
If X is negative or ends in a 0 and is not zero itself. We return false;

We store our reversedNumber in a long int to prevent a overflow when reversing our large number.

We create an int to hold our originalNumber x, to compare later.

While x does not equal 0. Extract the right most digit, update our reversedNumber to, reversedNumber * 10  +  (x % 10);
dividing the x by 10 to remove the right most digit.

If the originalNumber is equal to the reversedNumber return true, else return false.