mergeTwoSortedLists
First, I checked if either list is null. If there is we return the other list that isn't, since it is already sorted. Next, I created a pointer variable named head to keep track of the starting node, and set it to null. We compare the initial two nodes and start with the smallest of the two. In a loop, compared the value in list1 with the value in list2. If it was less than list2 we add them to the list, moving to the next node. If there are any remained nodes in the lists due to size difference. We add them to the list of sorted, this works because the remaining nodes are already sorted.