removeDuplicates.cpp
First, I checked if the nums array is empty, returning zero if was. Next, I created a new index variable initialized to zero to keep track of our position in our array where we will store the next unique value.  In a loop starting at index 1, we compare index 1 with index 0 checking if it is a duplicate. If its is not, we increment our new index variable, and store the unique value. Once the loop is done all unique elements are stored up to the new index + 1. We return the newIndex + 1 as the count.